Add list-containers access level to account ACLs in tempauth.
Supplementing the existing access levels "admin", "read-write", and
"read-only", the new "list-containers" access level allows GET and HEAD
operations on the account, but no operations on containers or objects.  This
allows users with this access to get a list of the containers in an account.
Further access within an individual container may be granted via container
ACLs, if desired.

Doc-Impact: Updated overview-auth.rst with the new access level.

Change-Id: If0c8d8d1037d595fad87122279b5ad0abf75331b